Switch back to web-mode for Typescript
Integration between various web languages including Javascript seems to be
more mature with web-mode, rather than having Typescript in its own mode as
well as Javascript and JSX, TSX, etc. I have also cleaned up the web-mode
config and set up a hook for LSP to be started with it.